# orchestrator/openapi.yaml
openapi: 3.1.0
info:
  title: Ops Orchestrator API
  version: 1.0.0
servers:
  - url: https://ops-orchestrator.onrender.com
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
paths:
...
  /ai/audit:
    get:
      summary: Read audit log
      security:
        - apiKeyAuth: []
      responses:
        '200': { description: OK }
    post:
      summary: Append an audit entry
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event: { type: string }
                data: { type: object }
      responses:
        '200': { description: OK }